refactor

+++ src/core/src/main/java/io/github/revNrun/revNrun/model/car/components/Brakes.java
@@ -6,7 +6,7 @@ import io.github.revNrun.revNrun.model.car.components.enums.CarSides;
 import java.util.List;
 
 public class Brakes extends WheelMountedComponent {
-    public Brakes(String name, float weight, int maxDurability, int currentDurability,
+    public Brakes(String name, float weight, float maxDurability, float currentDurability,
                   List<Effect> effects, CarAxis axle, CarSides side, float wearFactor) {
         super(name, weight, maxDurability, currentDurability, effects, axle, side, wearFactor);
     }
