+++ src/core/src/main/java/io/github/revNrun/revNrun/model/car/components/Sides.java
@@ -3,7 +3,7 @@ package io.github.revNrun.revNrun.model.car.components;
 import java.util.List;
 
 public class Sides extends AbstractComponent {
-    public Sides(String name, float weight, int maxDurability, int currentDurability, List<Effect> effects, float wearFactor) {
+    public Sides(String name, float weight, float maxDurability, float currentDurability, List<Effect> effects, float wearFactor) {
         super(name, weight, maxDurability, currentDurability, effects, wearFactor);
     }
 }
