+++ src/core/src/main/java/io/github/revNrun/revNrun/model/car/components/Tires.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public class Tires extends WheelMountedComponent {
 
-    public Tires(String name, float weight, int maxDurability, int currentDurability,
+    public Tires(String name, float weight, float maxDurability, float currentDurability,
                  List<Effect> effects, CarAxis axle, CarSides side, float wearFactor) {
         super(name, weight, maxDurability, currentDurability, effects, axle, side, wearFactor);
     }
