+++ src/core/src/main/java/io/github/revNrun/revNrun/model/car/components/WheelMountedComponent.java
@@ -9,7 +9,7 @@ public abstract class WheelMountedComponent extends AbstractComponent {
     protected CarAxis axle;
     protected CarSides side;
 
-    public WheelMountedComponent(String name, float weight, int maxDurability, int currentDurability,
+    public WheelMountedComponent(String name, float weight, float maxDurability, float currentDurability,
                                  List<Effect> effects, CarAxis axle, CarSides side, float wearFactor) {
         super(name, weight, maxDurability, currentDurability, effects, wearFactor);
         if (axle == null || side == null) {
